# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI MySQL Application",
    "description": "台股相關 **FastAPI + MySQL** 後端。\n\n- **互動文件**：本頁 Swagger UI（`/docs`）或 ReDoc（`/redoc`）。\n- **健康檢查**：`GET /health`。\n- **AI 分析**：見標籤「AI 分析」；分階架構可並行呼叫 `raw` 三筆 + `quick-insights` + `final`。\n\n實際部署網域與 CORS 請依環境調整。",
    "version": "1.0.0"
  },
  "paths": {
...
    "/stocks/{symbol}/latest": {
      "get": {
        "tags": [
          "股價查詢"
        ],
        "summary": "獲取最新股價",
        "description": "獲取指定股票的最新交易日價格數據。\n\n    **參數**：\n    - `symbol`: 股票代號（例如：2330、2317）\n\n    **返回**：\n    - 包含開盤價、最高價、最低價、收盤價、成交量等完整資訊",
        "operationId": "get_latest_price_stocks__symbol__latest_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回最新股價",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyPriceResponse"
                }
              }
            }
          },
          "404": {
            "description": "找不到該股票的數據"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stocks/{symbol}/history": {
      "get": {
        "tags": [
          "股價查詢"
        ],
        "summary": "Get Historical Prices",
        "description": "獲取指定股票的歷史價格數據",
        "operationId": "get_historical_prices_stocks__symbol__history_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "開始日期",
              "title": "Start Date"
            },
            "description": "開始日期"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "結束日期",
              "title": "End Date"
            },
            "description": "結束日期"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "跳過的記錄數",
              "default": 0,
              "title": "Skip"
            },
            "description": "跳過的記錄數"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "返回的記錄數",
              "default": 100,
              "title": "Limit"
            },
            "description": "返回的記錄數"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalPriceList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stocks/{symbol}/statistics": {
      "get": {
        "tags": [
          "股價查詢"
        ],
        "summary": "Get Statistics",
        "description": "獲取指定時間範圍的統計數據",
        "operationId": "get_statistics_stocks__symbol__statistics_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "開始日期",
              "title": "Start Date"
            },
            "description": "開始日期"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "結束日期",
              "title": "End Date"
            },
            "description": "結束日期"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceStatistics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stocks/compare/multiple": {
      "get": {
        "tags": [
          "股價查詢"
        ],
        "summary": "多股票價格比較",
        "description": "比較多支股票在同一時間範圍內的價格走勢。\n\n    **適用場景**：\n    - 繪製多條折線圖比較不同股票\n    - 分析股票間的相關性\n    - 投資組合績效對比\n\n    **參數**：\n    - `symbols`: 股票代號列表，用逗號分隔\n      - 範例：`2330,2317,2454`\n      - 最多支援 10 支股票同時比較\n    - `start_date`: 開始日期（YYYY-MM-DD）\n    - `end_date`: 結束日期（YYYY-MM-DD）\n\n    **返回數據格式**：\n    ```json\n    {\n      \"symbols\": [\"2330\", \"2317\"],\n      \"data\": [\n        {\n          \"date\": \"2024-03-10\",\n          \"prices\": {\n            \"2330\": 593.0,\n            \"2317\": 520.0\n          }\n        }\n      ]\n    }\n    ```\n\n    **注意**：\n    - 如某支股票在某日無交易，該日價格為 null\n    - 建議查詢時間不超過 6 個月以提升效能",
        "operationId": "compare_multiple_stocks_stocks_compare_multiple_get",
        "parameters": [
          {
            "name": "symbols",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "股票代號，用逗號分隔，例如: 2330,2317,2454",
              "title": "Symbols"
            },
            "description": "股票代號，用逗號分隔，例如: 2330,2317,2454"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "開始日期",
              "title": "Start Date"
            },
            "description": "開始日期"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "結束日期",
              "title": "End Date"
            },
            "description": "結束日期"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回多股票比較數據",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiStockResponse"
                }
              }
            }
          },
          "400": {
            "description": "參數錯誤（超過10支股票）"
          },
          "404": {
            "description": "找不到任何股票數據"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stocks/{symbol}/date-range": {
      "get": {
        "tags": [
          "股價查詢"
        ],
        "summary": "Get Symbol Date Range",
        "description": "獲取指定股票的日期範圍",
        "operationId": "get_symbol_date_range_stocks__symbol__date_range_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stocks/{symbol}/chart/candlestick-ma": {
      "get": {
        "tags": [
          "股價查詢"
        ],
        "summary": "K線圖 + 移動平均線",
        "description": "獲取K線圖數據並自動計算移動平均線（MA）。\n\n    **適用場景**：\n    - 專業股票技術分析\n    - 繪製 K線圖 + 多條 MA 線\n    - 成交量區間柱狀圖\n\n    **參數**：\n    - `symbol`: 股票代號（例如：2330）\n    - `start_date`: 開始日期（YYYY-MM-DD）\n    - `end_date`: 結束日期（YYYY-MM-DD）\n    - `ma_periods`: 移動平均線周期，用逗號分隔\n      - 預設：`5,10,20`\n      - 常用：`5,10,20,60,120`\n      - 最多支援 5 條 MA 線\n\n    **返回數據包含**：\n    - `candlestick`: 每日K線數據（OHLC + 成交量）\n    - `moving_averages`: 移動平均線數據\n      - MA5: 5日平均\n      - MA10: 10日平均\n      - MA20: 20日平均\n    - `dates`: 日期陣列（方便繪圖）\n\n    **使用範例**：\n    ```\n    GET /api/v1/stocks/2330/chart/candlestick-ma?start_date=2024-01-01&end_date=2024-03-10&ma_periods=5,10,20\n    ```\n\n    **前端圖表庫推薦**：\n    - ECharts（推薦）\n    - TradingView Lightweight Charts\n    - Chart.js with chartjs-chart-financial",
        "operationId": "get_candlestick_with_ma_stocks__symbol__chart_candlestick_ma_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "開始日期",
              "title": "Start Date"
            },
            "description": "開始日期"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "結束日期",
              "title": "End Date"
            },
            "description": "結束日期"
          },
          {
            "name": "ma_periods",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "移動平均線週期，用逗號分隔，例如: 5,10,20",
              "default": "5,10,20",
              "title": "Ma Periods"
            },
            "description": "移動平均線週期，用逗號分隔，例如: 5,10,20"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回K線 + MA 數據",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandlestickWithMAResponse"
                }
              }
            }
          },
          "400": {
            "description": "MA 周期格式錯誤或超過5條"
          },
          "404": {
            "description": "指定日期範圍內無數據"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DailyPriceResponse": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date",
            "description": "日期"
          },
          "symbol": {
            "type": "string",
            "maxLength": 10,
            "minLength": 1,
            "title": "Symbol",
            "description": "股票代號"
          },
          "open": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open",
            "description": "開盤價"
          },
          "high": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "High",
            "description": "最高價"
          },
          "low": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low",
            "description": "最低價"
          },
          "close": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close",
            "description": "收盤價"
          },
          "volume_shares": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Shares",
            "description": "成交股數"
          },
          "amount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount",
            "description": "成交金額"
          },
          "change": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change",
            "description": "漲跌價差"
          },
          "trades": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trades",
            "description": "成交筆數"
          }
        },
        "type": "object",
        "required": [
          "date",
          "symbol"
        ],
        "title": "DailyPriceResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}